feat(layout): use CSS grid instead of @lost for grid system
