+++ .stylelintrc.json
@@ -1,24 +1,18 @@
 {
-  "plugins": ["stylelint-prettier"],
   "extends": [
     "stylelint-config-standard",
     "stylelint-config-recess-order",
     "stylelint-config-prettier"
   ],
+  "plugins": ["stylelint-prettier"],
   "rules": {
+    "at-rule-no-unknown": null,
+    "no-descending-specificity": null,
     "prettier/prettier": true,
     "selector-pseudo-class-no-unknown": [
       true,
       {
         "ignorePseudoClasses": ["global"]
-      }
-    ],
-    "at-rule-no-unknown": null,
-    "no-descending-specificity": null,
-    "property-no-unknown": [
-      true,
-      {
-        "ignoreProperties": ["/lost-/"]
       }
     ]
   }
