+++ package.json
@@ -140,7 +140,6 @@
     "imagemin-mozjpeg": "^9.0.0",
     "imagemin-optipng": "^8.0.0",
     "jest": "^26.6.3",
-    "lost": "^8.3.1",
     "next-compose-plugins": "^2.2.1",
     "next-offline": "^5.0.3",
     "next-optimized-images": "^2.6.2",
@@ -149,8 +148,6 @@
     "node-fetch": "^2.6.1",
     "postcss-calc": "^7.0.5",
     "postcss-color-function": "^4.1.0",
-    "postcss-conditionals": "^2.1.0",
-    "postcss-each": "^0.10.0",
     "postcss-mixins": "^6.2.3",
     "postcss-preset-env": "^6.7.0",
     "prettier": "^2.2.1",
