+++ src/common/styles/layouts.css
@@ -1,0 +1,106 @@
+/* @styled-jsx=global */
+
+.l-container {
+  display: grid;
+  grid-template-areas: 'spacingLeft content spacingRight';
+  grid-template-columns: 1fr var(--layout-width) 1fr;
+  margin-right: 1rem;
+  margin-left: 1rem;
+
+  @media (--sm-up) {
+    grid-template-columns: 1fr var(--layout-width-sm-up) 1fr;
+  }
+
+  @media (--md-up) {
+    grid-template-columns: 1fr var(--layout-width-md-up) 1fr;
+  }
+
+  @media (--lg-up) {
+    grid-template-columns: 1fr var(--layout-width-lg-up) 1fr;
+  }
+
+  &.full {
+    margin-right: auto;
+    margin-left: auto;
+  }
+}
+
+/*
+ * Row
+ */
+.l-row {
+  display: grid;
+  grid-template-columns: repeat(4, [col-start] 1fr);
+  grid-column: content;
+  grid-column-gap: var(--gutter-width);
+
+  @media (--sm-up) {
+    grid-template-columns: repeat(8, [col-start] 1fr);
+  }
+
+  @media (--md-up) {
+    grid-template-columns: repeat(9, [col-start] 1fr);
+  }
+
+  @media (--lg-up) {
+    grid-template-columns: repeat(12, [col-start] 1fr);
+  }
+}
+
+/**
+ * Column
+ */
+.l-col-full {
+  grid-column: 1 / -1;
+}
+
+/*
+ * Three Column Layout
+ */
+.l-col-three-left {
+  grid-column: 1 / -1;
+
+  @media (--sm-up) {
+    grid-column: col-start 1 / span 1;
+  }
+
+  @media (--md-up) {
+    grid-column: col-start 1 / span 2;
+  }
+
+  @media (--lg-up) {
+    grid-column: col-start 1 / span 2;
+  }
+}
+
+.l-col-three-mid {
+  grid-column: 1 / -1;
+
+  @media (--sm-up) {
+    grid-column: col-start 2 / span 7;
+  }
+
+  @media (--md-up) {
+    grid-column: col-start 3 / span 7;
+  }
+
+  @media (--lg-up) {
+    grid-column: col-start 3 / span 7;
+  }
+}
+
+.l-col-three-right {
+  grid-column: 1 / -1;
+
+  @media (--sm-up) {
+    grid-column: col-start 2 / span 7;
+  }
+
+  @media (--md-up) {
+    grid-column: col-start 3 / span 7;
+  }
+
+  @media (--lg-up) {
+    grid-column: col-start 10 / span 3;
+  }
+}
