--- src/common/styles/layouts/grids.css
@@ -1,202 +1,0 @@
-/* @styled-jsx=global */
-
-@lost gutter 40px;
-@lost flexbox flex;
-@lost rounder 99.99999;
-
-/*
- * Row
- */
-.l-row {
-  lost-center: 100%;
-  margin-right: 1rem;
-  margin-left: 1rem;
-
-  @media (--sm-up) {
-    & {
-      lost-center: 704px;
-    }
-  }
-
-  @media (--md-up) {
-    & {
-      lost-center: 896px;
-    }
-  }
-
-  @media (--lg-up) {
-    & {
-      lost-center: 1208px;
-    }
-  }
-
-  &.full {
-    margin-right: auto;
-    margin-left: auto;
-  }
-}
-
-/*
- * Column & Offset
- */
-
-/* 4 cols in XSmall */
-@each $i in 0, 1, 2, 3, 4 {
-  .l-col-$(i) {
-    /* prettier-ignore */
-    lost-column: $(i)/4;
-
-    @if $(i) == 4 {
-      flex-basis: 100%;
-      width: 100%;
-      max-width: 100%;
-    }
-  }
-  .l-offset-$(i) {
-    /* prettier-ignore */
-    lost-offset: $(i)/4;
-  }
-}
-
-/* 8 cols in Small */
-@each $device in sm {
-  @media (--$(device)-up) {
-    @each $i in 0, 1, 2, 3, 4, 5, 6, 7, 8 {
-      .l-col-$(device)-$(i) {
-        /* prettier-ignore */
-        lost-column: $(i)/8;
-
-        @if $(i) == 8 {
-          flex-basis: 100%;
-          width: 100%;
-          max-width: 100%;
-        }
-      }
-      .l-offset-$(device)-$(i) {
-        /* prettier-ignore */
-        lost-offset: $(i)/8;
-      }
-    }
-  }
-}
-
-/* 9 cols in Medium */
-@each $device in md {
-  @media (--$(device)-up) {
-    @each $i in 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 {
-      .l-col-$(device)-$(i) {
-        /* prettier-ignore */
-        lost-column: $(i)/9;
-
-        @if $(i) == 9 {
-          flex-basis: 100%;
-          width: 100%;
-          max-width: 100%;
-        }
-      }
-      .l-offset-$(device)-$(i) {
-        /* prettier-ignore */
-        lost-offset: $(i)/9;
-      }
-    }
-  }
-}
-
-/* 12 cols in Large */
-@each $device in lg {
-  @media (--$(device)-up) {
-    @each $i in 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12 {
-      .l-col-$(device)-$(i) {
-        /* prettier-ignore */
-        lost-column: $(i)/12;
-
-        @if $(i) == 12 {
-          flex-basis: 100%;
-          width: 100%;
-          max-width: 100%;
-        }
-      }
-      .l-offset-$(device)-$(i) {
-        /* prettier-ignore */
-        lost-offset: $(i)/12;
-      }
-    }
-  }
-}
-
-/*
- * Three Column Layout
- */
-.l-col-three-left {
-  lost-column: 4/4;
-
-  @media (--sm-up) {
-    & {
-      lost-column: 1/8;
-    }
-  }
-
-  @media (--md-up) {
-    & {
-      lost-column: 2/9;
-    }
-  }
-
-  @media (--lg-up) {
-    & {
-      lost-column: 2/12;
-    }
-  }
-}
-
-.l-col-three-mid {
-  lost-column: 4/4;
-
-  @media (--sm-up) {
-    & {
-      lost-column: 7/8;
-    }
-  }
-
-  @media (--md-up) {
-    & {
-      lost-column: 7/9;
-    }
-  }
-
-  @media (--lg-up) {
-    & {
-      lost-column: 7/12;
-    }
-  }
-
-  @media (--lg-down) {
-    margin-right: 0 !important;
-  }
-}
-
-.l-col-three-right {
-  lost-column: 4/4;
-
-  @media (--sm-up) {
-    & {
-      lost-offset: 1/8;
-      lost-column: 7/8;
-    }
-  }
-
-  @media (--md-up) {
-    & {
-      lost-offset: 2/9;
-      lost-column: 7/9;
-    }
-  }
-
-  @media (--lg-up) {
-    & {
-      lost-offset: 0/12;
-      lost-column: 3/12;
-      margin-right: 0 !important;
-    }
-  }
-}
