+++ src/components/Dialog/styles.css
@@ -47,6 +47,34 @@
       height: auto;
     }
   }
+
+  &.sm {
+    grid-column: 1 / -1;
+
+    @media (--sm-up) {
+      grid-column: col-start 3 / span 4;
+    }
+
+    @media (--lg-up) {
+      grid-column: col-start 5 / span 4;
+    }
+  }
+
+  &.lg {
+    grid-column: 1 / -1;
+
+    @media (--sm-up) {
+      grid-column: col-start 2 / span 6;
+    }
+
+    @media (--md-up) {
+      grid-column: col-start 3 / span 5;
+    }
+
+    @media (--lg-up) {
+      grid-column: col-start 4 / span 6;
+    }
+  }
 }
 
 .handle {
