+++ src/components/Layout/FixedMain/styles.css
@@ -5,6 +5,24 @@
   pointer-events: none;
 }
 
-.inner {
+.single-col {
+  & .content {
+    grid-column: 1 / -1;
+
+    @media (--sm-up) {
+      grid-column: col-start 2 / span 6;
+    }
+
+    @media (--md-up) {
+      grid-column: col-start 3 / span 5;
+    }
+
+    @media (--lg-up) {
+      grid-column: col-start 4 / span 6;
+    }
+  }
+}
+
+.content {
   pointer-events: initial;
 }
