+++ src/views/Migration/Banner/styles.css
@@ -1,0 +1,28 @@
+.banner {
+  padding: var(--spacing-base);
+  margin-top: 5rem;
+  margin-bottom: 8rem;
+
+  & h4 {
+    font-size: var(--font-size-md);
+    color: var(--color-matters-green);
+  }
+
+  & p {
+    margin: var(--spacing-x-tight) 0 var(--spacing-base);
+    font-size: var(--font-size-sm);
+    color: var(--color-grey-darker);
+  }
+
+  @media (--sm-up) {
+    padding: var(--spacing-x-loose) var(--spacing-xx-loose);
+
+    & h4 {
+      font-size: var(--font-size-lg);
+    }
+
+    & p {
+      font-size: var(--font-size-md);
+    }
+  }
+}
