+++ src/views/Migration/Features/styles.css
@@ -1,0 +1,76 @@
+.features {
+  margin-top: 5rem;
+}
+
+h2 {
+  padding-bottom: var(--spacing-x-tight);
+  text-align: center;
+}
+
+ul {
+  & li {
+    margin-top: var(--spacing-xx-loose);
+  }
+
+  @media (--sm-up) {
+    & li {
+      margin-top: var(--spacing-xxx-loose);
+    }
+  }
+}
+
+.content {
+  align-self: center;
+
+  & h4 {
+    font-size: var(--font-size-xl);
+    font-weight: var(--font-weight-semibold);
+    color: var(--color-matters-green);
+  }
+
+  & p {
+    margin: var(--spacing-base) 0;
+    font-size: var(--font-size-sm);
+    color: var(--color-grey-darker);
+  }
+
+  & a {
+    font-size: var(--font-size-sm);
+  }
+}
+
+img {
+  justify-self: center;
+}
+
+.content,
+img {
+  grid-column: 1 / -1;
+
+  @media (--sm-up) {
+    grid-column: col-start 2 / span 6;
+  }
+
+  @media (--md-up) {
+    grid-column: col-start 2 / span 7;
+  }
+}
+
+@media (--lg-up) {
+  .content {
+    grid-column: col-start 1 / span 5;
+  }
+  img {
+    grid-column: col-start 7 / span 6;
+  }
+
+  li:nth-child(2) {
+    & .content {
+      grid-column: col-start 8 / span 5;
+    }
+    & img {
+      grid-column: col-start 1 / span 6;
+      order: -1;
+    }
+  }
+}
