+++ src/views/Migration/Intro/styles.css
@@ -1,0 +1,14 @@
+.intro {
+  @mixin flex-center-all;
+
+  height: 100vh;
+  height: calc(var(--vh) * 100);
+  padding-top: var(--global-header-height);
+  text-align: center;
+  background-color: var(--color-green-lighter);
+
+  & p {
+    padding: var(--spacing-x-loose) 0;
+    color: var(--color-grey-darker);
+  }
+}
