+++ src/views/Migration/styles.css
@@ -1,11 +1,8 @@
-h2 {
-  @mixin font-serif;
+main {
+  & :global(h2) {
+    @mixin font-serif;
 
-  font-size: 1.7rem;
-  line-height: 1.4;
-
-  @media (--sm-up) {
-    font-size: 2.5rem;
+    font-size: var(--font-size-xxl);
   }
 }
 
@@ -15,204 +12,12 @@ header {
   right: 0;
   left: 0;
   z-index: 1;
-  padding: var(--spacing-x-loose) 0;
+  padding: var(--spacing-xx-loose) 0;
 
   & .logo {
     & :global(svg) {
       width: 97px;
       height: 20px;
-    }
-  }
-}
-
-.intro-wrap {
-  @mixin flex-center-all;
-
-  height: 100vh;
-  height: calc(var(--vh) * 100);
-  padding-top: var(--global-header-height);
-  background-color: var(--color-green-lighter);
-  background-repeat: no-repeat;
-
-  @media (--sm-up) {
-    height: 50rem;
-    padding-top: 0;
-  }
-}
-
-.intro {
-  @mixin flex-center-all;
-
-  width: 100%;
-  text-align: center;
-
-  & section {
-    max-width: 43rem;
-
-    & p {
-      padding: var(--spacing-x-loose) 0;
-      color: var(--color-grey-darker);
-    }
-  }
-}
-
-.steps-wrap {
-  padding-top: 8rem;
-
-  & .header {
-    @mixin flex-center-all;
-  }
-
-  & .steps {
-    @mixin flex-start-center;
-
-    flex-direction: column;
-
-    @media (--lg-up) {
-      @mixin flex-start-space-between;
-
-      flex-direction: row;
-    }
-  }
-}
-
-.step {
-  @mixin flex-center-all;
-
-  flex-direction: column;
-  width: 100%;
-  margin-top: var(--spacing-xx-loose);
-
-  @media (--lg-up) {
-    max-width: 20.5rem;
-    margin-top: var(--spacing-loose);
-  }
-
-  & .avatar {
-    display: flex;
-    align-items: center;
-    padding: var(--spacing-xx-tight);
-    padding-right: var(--spacing-tight);
-    font-size: var(--color-black);
-    background: var(--color-grey-lighter);
-    border-radius: 2rem;
-
-    & .info {
-      margin-left: var(--spacing-x-tight);
-    }
-  }
-
-  & .title {
-    font-size: var(--font-size-xm);
-    font-weight: var(--font-weight-medium);
-
-    & span.number {
-      padding-right: var(--spacing-x-tight);
-      color: var(--color-grey-light);
-    }
-  }
-
-  & .content,
-  & .sub-content {
-    margin-top: var(--spacing-base);
-    margin-bottom: var(--spacing-x-tight);
-    font-size: var(--font-size-sm);
-    color: var(--color-grey-dark);
-  }
-
-  & .content {
-    max-width: 16.5rem;
-    margin-bottom: var(--spacing-loose);
-    text-align: center;
-  }
-
-  & .sub-content-link {
-    font-size: var(--font-size-sm);
-    color: var(--color-matters-green);
-  }
-}
-
-.features-wrap {
-  margin-top: 6rem;
-
-  & .header {
-    @mixin flex-center-all;
-
-    text-align: center;
-  }
-
-  & .features {
-    margin-top: var(--spacing-xx-loose);
-
-    @media (--lg-up) {
-      margin-top: 4.24rem;
-    }
-  }
-}
-
-.feature {
-  @mixin flex-center-all;
-
-  flex-direction: column;
-  margin-top: var(--spacing-loose);
-
-  & section {
-    width: 100%;
-  }
-
-  & .title {
-    font-size: 1.625rem;
-    font-weight: var(--font-weight-semibold);
-    color: var(--color-matters-green);
-  }
-
-  & .content {
-    margin: var(--spacing-base) 0;
-    font-size: var(--font-size-sm);
-    color: var(--color-grey-darker);
-  }
-
-  & .link {
-    font-size: var(--font-size-sm);
-  }
-}
-
-.banner {
-  @mixin flex-center-all;
-
-  flex-direction: column;
-  min-height: 10.75rem;
-  padding: var(--spacing-tight);
-  margin-top: 5rem;
-  margin-bottom: 8rem;
-
-  & p,
-  & section {
-    width: 100%;
-  }
-
-  & .title {
-    font-size: var(--font-size-md);
-    font-weight: var(--font-weight-semibold);
-    line-height: 1.2;
-    color: var(--color-matters-green);
-  }
-
-  & .content {
-    margin: var(--spacing-x-tight) 0 var(--spacing-base);
-    font-size: var(--font-size-sm);
-    color: var(--color-grey-darker);
-  }
-
-  @media (--sm-up) {
-    padding: 0 var(--spacing-xx-loose);
-
-    & .title {
-      font-size: var(--font-size-lg);
-    }
-
-    & .content {
-      font-size: var(--font-size-md);
     }
   }
 }
